Again with the compression pipeline errors.

+++ pipelines/changeo-clone.sh
@@ -177,8 +177,8 @@ if $FUNCTIONAL; then
         --outname "${OUTNAME}" --outdir . \
         >> $PIPELINE_LOG 2> $ERROR_LOG
     check_error
-    SELECT_FILE="${OUTNAME}_parse-select.${EXT}"
-    LAST_FILE=$SELECT_FILE
+    SELECT_PASS="${OUTNAME}_parse-select.${EXT}"
+    LAST_FILE=$SELECT_PASS
 else
     LAST_FILE=${DB}
 fi
@@ -190,8 +190,8 @@ DefineClones.py ${DC_COMMAND} -d ${LAST_FILE} --model ${DC_MODEL} \
     --outname "${OUTNAME}" --outdir . --log "${LOGDIR}/clone.log" \
     >> $PIPELINE_LOG 2> $ERROR_LOG
 check_error
-CLONE_FILE="${OUTNAME}_clone-pass.${EXT}"
-LAST_FILE=$CLONE_FILE
+CLONE_PASS="${OUTNAME}_clone-pass.${EXT}"
+LAST_FILE=$CLONE_PASS
 
 # Create germlines
 if $GERMLINES; then
@@ -200,8 +200,8 @@ if $GERMLINES; then
         --sf ${CG_SFIELD} --vf ${CG_VFIELD} --cloned --outname "${OUTNAME}" \
         >> $PIPELINE_LOG 2> $ERROR_LOG
 	check_error
-	GERM_FILE="${OUTNAME}_germ-pass.${EXT}"
-	LAST_FILE=$GERM_FILE
+	GERM_PASS="${OUTNAME}_germ-pass.${EXT}"
+	LAST_FILE=$GERM_PASS
 fi
 
 # Process log files
@@ -210,17 +210,26 @@ ParseLog.py -l "${LOGDIR}/clone.log" -f VALLELE DALLELE JALLELE JUNCLEN SEQUENCE
     > /dev/null 2> $ERROR_LOG &
 wait
 
-# Zip or delete intermediate and log files
+# Zip or delete log files
 printf "  %2d: %-*s $(date +'%H:%M %D')\n" $((++STEP)) 24 "Compressing files"
 LOG_FILES=$(ls ${LOGDIR}/*.log | grep -v "pipeline")
-TEMP_FILES=$(ls ${SELECT_FILE} ${CLONE_FILE} ${GERM_FILE}  2>/dev/null | grep -v "${LAST_FILE}\|$(basename ${DB})")
-if $ZIP_FILES; then
-    tar -zcf log_files.tar.gz $LOG_FILES
-    tar -zcf temp_files.tar.gz $TEMP_FILES
+if [[ ! -z $LOG_FILES ]]; then
+    if $ZIP_FILES; then
+        tar -zcf log_files.tar.gz $LOG_FILES
+    fi
+    if $DELETE_FILES; then
+        rm $LOG_FILES
+    fi
 fi
-if $DELETE_FILES; then
-    rm $TEMP_FILES
-    rm $LOG_FILES
+# Zip or delete intermediate files
+TEMP_FILES=$(ls ${SELECT_PASS} ${CLONE_PASS} ${GERM_PASS}  2>/dev/null | grep -v "${LAST_FILE}\|$(basename ${DB})")
+if [[ ! -z $TEMP_FILES ]]; then
+    if $ZIP_FILES; then
+        tar -zcf temp_files.tar.gz $TEMP_FILES
+    fi
+    if $DELETE_FILES; then
+        rm $TEMP_FILES
+    fi
 fi
 
 # End
